making SSL mount optional for AdminApp

+++ Web-Ods-AdminApp/Alpine/mssql/run.sh
@@ -31,7 +31,9 @@ mv /app/temp.json /app/appsettings.json
 # >&2 echo "Postgres is up - executing command"
 # exec $cmd
 
-cp /ssl/server.crt /usr/local/share/ca-certificates/
-update-ca-certificates
+if [ -f /ssl/server/crt] then
+  cp /ssl/server.crt /usr/local/share/ca-certificates/
+  update-ca-certificates
+fi
 
 dotnet EdFi.Ods.AdminApp.Web.dll
